Update create_timepoint.class.inc
Added the missing MCI visits which were causing the problem of timepoints being unable to be created from the front-end by end-users.

This whole hack needs to be replaced by a proper call to the config.xml file using the appropriate Config class of Loris. For now, I re-ordered the declaration of the array so that it can be maintained more easily. Still, needs to be refactored.