Bugfix candidate list access profile panel webpack entry
The file for the access profile panel wasn't included in the webpack entries despite having a compiled version in the js folder. Added the file to the entry object using an array for compatibility with the use-case.

Initial commit for new public acknowledgement page

Initial commit of a new version of the public facing acknowledgements page based on a better DB modeling, added error handling and new-ish design and design elements.

Correction of terminology used

The study is PREVENT-AD as in the acronym for PRe-symptomatic EValuation of Experimental or Novel Treatments for Alzheimer’s Disease, not a proper noun.

Update acknowledgements2.php

Adding support for the display of present and past contributors and support for the role ordering.

Update acknowledgements.css

Adding stylings for the past contributors.

Update acknowledgements2.tpl

Correcting the misplaced TR tag and adding a separate loop to process the past contributors.

Update acknowledgements2.tpl

Cleaning template by removing repeated code

Update acknowledgements2.tpl

Spacing adjustments.

Update acknowledgements.css

Added padding at the bottom of the page.

Update acknowledgements2.tpl

Testing the past contributors array for emptyness to avoid displaying the separation without any past contributors.

Update User.class.inc

Basic update to avoid undefined behavior when encountering the empty results scenario.

Update acknowledgements2.tpl

Changing the text for the hyperlinks and the separator element.

Adding download buttons for both an XML and a TSV file.

Update acknowledgements.css

Adding stylings for the separator element and the download buttons

Update acknowledgements2.php

Adding the download feature for both XML and TSV.

Adding support for timed roles while leaving the default as always on if neither the start or end dates are defined.

Update acknowledgements2.php

Added comments in last functions

Replacing the old version of the PHP

Replacing the old version of the PHP file with the new one.

Replacing the old version of the template

Replacing the old version of the template with the new one

Update acknowledgements_errors.tpl

Adapting the text of the error template to reflect the new address of the pages

Updating public facing acknowledgements

Adding support for a value to the authors flag to allow for the "preview" link from the acknowledgement module using a checkbox.

Bugfix to the candidate list access profile panel

A missing slash in the redirect command from the profile panel made the link redirect to the dashboard instead of the candidate page. Adding the correction and the compiled versions through webpack.

Bugfix to remove warning and potential problems

Testing for the existence of variables before using them. In this case the properties subtest and submenu from the REQUEST super global.