Update to the SART parsing script
The V2 and V3 edat files produce a column more than the v1. Thus, checking by the size of the header array doesn't give consistent results. Replaced that check with a confirmation that all calculated columns are present. Combined with the check on length, it should work for all current versions and account for possible changes in further versions of the edat files.