New HTML notification template to be used to pass contacts from the contact module
Smarty template implementing a generic HTML email format to send the information from a contact of the contact module. It maintains compatibility with the extraction of the subject from the topline, as was done in previous text-only templates, but it adds better flexibility to make the output less jarring when viewed with typical modern email clients.