New template for new account request notifications using a responsive HTML format
New smarty template using a generic responsive HTML format to send the information needed to evaluate a new account request to the administrator. It maintains the first line extraction of the subject for legacy purposes but displays the information in a manner less jarring than text-only output in a modern email client.